Fix the build when using OptiX 5.1.0 and make DevIL dependency optional.
OptiX 5.1.0 changed the OptiX core library names to contain the major and minor version numbers.
FindOptiX.cmake now determines the name by parsing the OptiX_INSTALL_DIR CMake variable which always needs to be set.

Additionally the DevIL image library dependency has been made optional with a warning.
If the CMake variable IL_FOUND is not true, projects depending on DevIL will not be included in the solution and the dynamic link libraries will not be copied into the build target folder.